my commit

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,15 +1,19 @@
 #include "lists.h"
 #include <stdio.h>
-
+/**
+ * listint_len - returns the number of elements in a linked
+ * @h: head of list
+ * Return: number of elements
+*/
 
 
 size_t listint_len(const listint_t *h)
 {
-long unsigned int k = 0;
+int k = 0;
 while (h != NULL)
 {
 h = h->next;
 k++;
 }
-return(k);
+return (k);
 }
